fix: correct SQL placeholders and JWT middleware for SQLite compatibility

+++ internal/domain/application.go
@@ -3,15 +3,15 @@ package domain
 import "time"
 
 type Application struct {
-	ID         int64
+	ID         int
 	Url        string
 	Country    string
-	Created_by int64
+	Created_by int
 	Updated_at time.Time
 	Created_at time.Time
 }
 
-func NewApplication(url string, country string, created_by int64) (*Application, error) {
+func NewApplication(url string, country string, created_by int) (*Application, error) {
 	application := Application{
 		Url:        url,
 		Country:    country,
