+++ internal/repository/contracts.go
@@ -17,20 +17,20 @@ func NewSQLiteRepository(db *sql.DB) (UserRepository, ApplicationRepository) {
 
 type UserRepository interface {
 	CreateUser(user *domain.User) error
-	FindUserByID(id int64) (*domain.User, error)
+	FindUserByID(id int) (*domain.User, error)
 	FindUserByEmail(email string) (*domain.User, error)
 	FindUserByUsername(username string) (*domain.User, error)
 	ListUsers() ([]*domain.User, error)
 	UpdateUser(user *domain.User) error
-	DeleteUserByID(id int64) error
+	DeleteUserByID(id int) error
 }
 
 type ApplicationRepository interface {
 	CreateApplication(application *domain.Application) error
-	FindApplicationByID(id int64) (*domain.Application, error)
+	FindApplicationByID(id int) (*domain.Application, error)
 	FindApplicationByURL(url string) (*domain.Application, error)
 	FindApplicationByCountry(country string) (*domain.Application, error)
 	ListApplications() ([]*domain.Application, error)
 	UpdateApplication(application *domain.Application) error
-	DeleteApplicationByID(id int64) error
+	DeleteApplicationByID(id int) error
 }
